..F....... [ZBX-14663] fixed calculation of Y zero position in graph

+++ frontends/php/include/classes/graphdraw/CLineGraphDraw.php
@@ -757,7 +757,7 @@ class CLineGraphDraw extends CGraphDraw {
 				}
 			}
 			else {
-				$this->zero[$side] = $this->sizeY + $this->shiftY - (int) abs($this->m_minY[$side] / $this->unit2px[$side]);
+				$this->zero[$side] = $this->sizeY + $this->shiftY - (float) abs($this->m_minY[$side] / $this->unit2px[$side]);
 				$this->oxy[$side] = 0;
 			}
 		}
